+++ database/migrations/2017_05_09_174758_create_table_dir_global.php
@@ -13,17 +13,18 @@ class CreateTableDirGlobal extends Migration
      */
     public function up()
     {
-        Schema::create('dir_global', function (Blueprint $table) {
-            $table->increments('id');
-            $table->string('dir_type_id');
-            $table->string('caption');
-            $table->string('group_name')->nullable(true);
-            $table->integer('order')->nullable(true);
-            $table->boolean('is_deleted')->default(false);
-            $table->timestamps();
-            $table->foreign('dir_type_id')->references('id')->on('dir_types')->onDelete('cascade')->onUpdate('cascade');
-
-        });
+        if (!Schema::hasTable('dir_global')) {
+            Schema::create('dir_global', function (Blueprint $table) {
+                $table->increments('id');
+                $table->string('dir_type_id');
+                $table->string('caption');
+                $table->string('group_name')->nullable(true);
+                $table->integer('order')->nullable(true);
+                $table->boolean('is_deleted')->default(false);
+                $table->timestamps();
+                $table->foreign('dir_type_id')->references('id')->on('dir_types')->onDelete('cascade')->onUpdate('cascade');
+            });
+        }
     }
 
     /**
@@ -33,6 +34,6 @@ class CreateTableDirGlobal extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('dir_global');
+//        Schema::dropIfExists('dir_global');
     }
 }
